pyproject.toml
[project]
name = "mywbooks"
version = "0.1.0"
description = "The WBooks can manage your collection of web books. Download new chapters, convert to epub and send it to your reading device"
readme = "README.md"
authors = [
    { name = "Nokorot", email = "torhaakon2012@hotmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "bs4>=0.0.2",
    "ebooklib>=0.19",
    "fastapi[standard]>=0.115.14",
    "mypy>=1.18.1",
    "pillow>=11.3.0",
    "sqlalchemy>=2.0.43",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mywbooks",
]

[tool.uv.sources]
mywbooks = { workspace = true }
